cleaner file deletion
on notebook deletion, check if folder to delete exist to avoid exception
in some cases (nose tests)

+++ converters/base.py
@@ -72,11 +72,11 @@ class Converter(object):
         if not os.path.isdir(files_dir):
             os.mkdir(files_dir)
         self.infile_root = infile_root
-        self.files_dir = files_dir
+        self.files_dir = os.path.abspath(files_dir)
         self.outbase = os.path.join(self.infile_dir, infile_root)
 
     def __del__(self):
-        if not os.listdir(self.files_dir):
+        if os.path.isdir(self.files_dir) and not os.listdir(self.files_dir):
             os.rmdir(self.files_dir)
 
     def dispatch(self, cell_type):
